#104 refactor: pretendard 폰트 설치 및 console.log 정리
