+++ package.json
@@ -28,6 +28,7 @@
     "framer-motion": "^11.3.8",
     "gsap": "^3.12.5",
     "jotai": "^2.9.2",
+    "pretendard": "^1.3.9",
     "react": "^18.3.1",
     "react-autosize-textarea": "^7.1.0",
     "react-beautiful-dnd": "^13.1.1",
